fix relation usecase

+++ entities/models/credit.go
@@ -6,7 +6,7 @@ type Credit struct {
 	ID          uint64
 	FromUserID  uint64 //yg kasik credit
 	ToUserID    uint64 //yg nerima credit
-	Credit      float64
+	Credit      float32
 	Description string
 	CreatedAt   time.Time
 	UpdatedAt   time.Time
